fixing error strings-to-time conversion
Someone introduced a bug in which a week would be added because -- for some reason -- the person included the working hours in work days in the the conversion. Seven years later, I am trying to quickly make some Gantt charts with effort estimates, and I was very confused by this bug.